backend/requirements.txt
fastapi==0.95.2
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==1.10.13
passlib==1.7.4
python-multipart==0.0.6
python-jose==3.3.0
psycopg2-binary 